requirements.txt
tqdm
pytest
Pillow
torch
torchvision
tensorboard
wandb
numpy
filelock
